+++ internal/filesystem/enabled.go
@@ -16,7 +16,7 @@ import (
 )
 
 // Open opens a file.
-func (fs *Fuse) Open(path string, flags int) (errc int, fh uint64) {
+func (fs *Fuse) Open(path string, _ int) (errc int, fh uint64) {
 	defer fs.synchronize()()
 	logs.Debug("Opening file ", filepath.FromSlash(path))
 
@@ -45,7 +45,7 @@ func (fs *Fuse) Open(path string, flags int) (errc int, fh uint64) {
 			return -fuse.EIO, ^uint64(0)
 
 		} else if n.node.stat.Size != newSize {
-			fs.updateNodeSizesAlongPath(path, n.node.stat.Size-newSize, fuse.Now())
+			fs.updateNodeSizesAlongPath(path, newSize-n.node.stat.Size, fuse.Now())
 		}
 		n.node.decryptionChecked = true
 	}
@@ -164,7 +164,7 @@ func (fs *Fuse) Read(path string, buff []byte, ofst int64, fh uint64) int {
 
 // Readdir reads the contents of a directory.
 func (fs *Fuse) Readdir(path string, fill func(name string, stat *fuse.Stat_t, ofst int64) bool,
-	ofst int64, fh uint64) (errc int) {
+	_ int64, fh uint64) (errc int) {
 	defer fs.synchronize()()
 	node := fs.getNode(path, fh).node
 	if node == nil {